Add unix-dgram support and es6 syntax.
