--- .travis.yml
@@ -1,13 +1,0 @@
-language: node_js
-
-node_js:
-  - "6"
-  - "8"
-  - "10"
-  - "node"
-
-install:
-  - npm install
-
-script:
-  - npm run-script coverage
