+++ lib/helpers.js
@@ -106,9 +106,25 @@ function getDefaultRoute() {
   return null;
 }
 
+/**
+ * High-resolution timer
+ */
+function createHrTimer() {
+  const start = process.hrtime();
+
+  return () => {
+    const durationComponents = process.hrtime(start);
+    const seconds = durationComponents[0];
+    const nanoseconds = durationComponents[1];
+    const duration = (seconds * 1000) + (nanoseconds / 1E6);
+    return duration;
+  };
+}
+
 module.exports = {
-  formatTags: formatTags,
-  overrideTags: overrideTags,
-  formatDate: formatDate,
-  getDefaultRoute: getDefaultRoute
+  formatTags,
+  overrideTags,
+  formatDate,
+  getDefaultRoute,
+  createHrTimer
 };
