+++ lib/statsd.js
@@ -1,492 +1,337 @@
-const dgram = require('dgram'),
-  util = require('util'),
-  dns = require('dns'),
-  net = require('net'),
-  helpers = require('./helpers'),
-  applyStatsFns = require('./statsFunctions');
+const Transport = require('./transport');
+const helpers = require('./helpers');
 
 /**
- * The Client for StatsD.  The main entry-point for hot-shots.  Note adding new parameters
- * to the constructor is deprecated- please use the constructor as one options object.
- * @constructor
+ * StatsD Client
+ * @description The main entry-point for hot-shots.
  */
-const Client = function (host, port, prefix, suffix, globalize, cacheDns, mock,
-    globalTags, maxBufferSize, bufferFlushInterval, telegraf, sampleRate, protocol) {
-  let options = host || {};
-  const self = this;
+class Client extends Transport {
 
-  // Adding options below is DEPRECATED.  Use the options object instead.
-  if (arguments.length > 1 || typeof(host) === 'string') {
-    options = {
-      host        : host,
-      port        : port,
-      prefix      : prefix,
-      suffix      : suffix,
-      globalize   : globalize,
-      cacheDns    : cacheDns,
-      mock        : mock === true,
-      globalTags  : globalTags,
-      maxBufferSize : maxBufferSize,
-      bufferFlushInterval: bufferFlushInterval,
-      telegraf    : telegraf,
-      sampleRate  : sampleRate,
-      protocol    : protocol
-    };
-  }
-
-  const createSocket = function createSocket(instance, args) {
-    let socket;
-    let errMessage;
-
-    if (args.protocol === 'tcp') {
-      try {
-        socket = net.connect(args.port, args.host);
-        socket.setKeepAlive(true);
-      } catch (e) {
-        errMessage = `Could not establish connection to ${args.host}:${args.port}`;
-        if (instance.errorHandler) {
-          instance.errorHandler(new Error(errMessage));
-        } else {
-          console.log(errMessage);
-        }
-      }
-    } else {
-      socket = dgram.createSocket('udp4');
-    }
-
-    return socket;
-  };
-
-  // hidden global_tags option for backwards compatibility
-  options.globalTags = options.globalTags || options.global_tags;
-
-  this.protocol    = (options.protocol && options.protocol.toLowerCase());
-  this.host        = options.host || 'localhost';
-  this.port        = options.port || 8125;
-  this.prefix      = options.prefix || '';
-  this.suffix      = options.suffix || '';
-  this.socket      = options.isChild ? options.socket : createSocket(this, {
-    host: this.host,
-    port: this.port,
-    protocol: this.protocol
-  });
-  this.mock        = options.mock;
-  this.globalTags  = typeof options.globalTags === 'object' ?
-      helpers.formatTags(options.globalTags, options.telegraf) : [];
-  this.telegraf    = options.telegraf || false;
-  this.maxBufferSize = options.maxBufferSize || 0;
-  this.sampleRate  = options.sampleRate || 1;
-  this.bufferFlushInterval = options.bufferFlushInterval || 1000;
-  this.bufferHolder = options.isChild ? options.bufferHolder : { buffer: '' };
-  this.errorHandler = options.errorHandler;
-
-  // If we're mocking the client, create a buffer to record the outgoing calls.
-  if (this.mock) {
-    this.mockBuffer = [];
-  }
-
-  // We only want a single flush event per parent and all its child clients
-  if (!options.isChild && this.maxBufferSize > 0) {
-    this.intervalHandle = setInterval(this.onBufferFlushInterval.bind(this), this.bufferFlushInterval);
-  }
-
-  if (options.isChild) {
-    if (options.dnsError) {
-      this.dnsError = options.dnsError;
-    }
-  } else if (options.cacheDns === true) {
-    dns.lookup(options.host, (err, address) => {
-      if (err === null) {
-        self.host = address;
-      } else {
-        self.dnsError = err;
-      }
-    });
-  }
-
-  if (!options.isChild && options.errorHandler) {
-    this.socket.on('error', options.errorHandler);
-  }
-
-  if (options.globalize) {
-    global.statsd = this;
-  }
-
-  if (options.useDefaultRoute) {
-    const defaultRoute = helpers.getDefaultRoute();
-    if (defaultRoute) {
-      console.log(`Got ${defaultRoute} for the system's default route`);
-      this.host = defaultRoute;
-    }
-  }
-
-  this.messagesInFlight = 0;
-  this.CHECKS = {
-    OK: 0,
-    WARNING: 1,
-    CRITICAL: 2,
-    UNKNOWN: 3,
-  };
-};
-
-applyStatsFns(Client);
-
-/**
- * Checks if stats is an array and sends all stats calling back once all have sent
- * @param stat {String|Array} The stat(s) to send
- * @param value The value to send
- * @param type The type of the metric
- * @param sampleRate {Number=} The Number of times to sample (0 to 1). Optional.
- * @param tags {Array=} The Array of tags to add to metrics. Optional.
- * @param callback {Function=} Callback when message is done being delivered. Optional.
- */
-Client.prototype.sendAll = function (stat, value, type, sampleRate, tags, callback) {
-  let completed = 0;
-  let calledback = false;
-  let sentBytes = 0;
-  const self = this;
-
-  if (sampleRate && typeof sampleRate !== 'number') {
-    callback = tags;
-    tags = sampleRate;
-    sampleRate = undefined;
-  }
-
-  if (tags && typeof tags !== 'object') {
-    callback = tags;
-    tags = undefined;
+  /**
+   * Represents the timing stat
+   * @param stat {String|Array} The stat(s) to send
+   * @param time {Number} The time in milliseconds to send
+   * @param sampleRate {Number=} The Number of times to sample (0 to 1). Optional.
+   * @param tags {Array=} The Array of tags to add to metrics. Optional.
+   * @param callback {Function=} Callback when message is done being delivered. Optional.
+   */
+  timing(stat, time, sampleRate, tags, callback) {
+    this.sendAll(stat, time, 'ms', sampleRate, tags, callback);
   }
 
   /**
-   * Gets called once for each callback, when all callbacks return we will
-   * call back from the function
-   * @private
+   * Represents the timing stat by recording the duration a function takes to run (in milliseconds)
+   * @param func {Function} The function to run
+   * @param stat {String|Array} The stat(s) to send
+   * @param sampleRate {Number=} The Number of times to sample (0 to 1). Optional.
+   * @param tags {Array=} The Array of tags to add to metrics. Optional.
+   * @param callback {Function=} Callback when message is done being delivered. Optional.
    */
-  function onSend(error, bytes) {
-    completed += 1;
-    if (calledback) {
-      return;
-    }
+  timer(func, stat, sampleRate, tags, callback) {
+    return (...args) => {
+      const start = process.hrtime();
+      try {
+        return func(...args);
+      } finally {
+        // get duration in milliseconds
+        const durationComponents = process.hrtime(start);
+        const seconds = durationComponents[0];
+        const nanoseconds = durationComponents[1];
+        const duration = (seconds * 1000) + (nanoseconds / 1E6);
 
-    if (error) {
-      if (typeof callback === 'function') {
-        calledback = true;
-        callback(error);
-      } else if (self.errorHandler) {
-        calledback = true;
-        self.errorHandler(error);
+        this.timing(
+          stat,
+          duration,
+          sampleRate,
+          tags,
+          callback
+        );
       }
-      return;
-    }
-
-    if (bytes) {
-      sentBytes += bytes;
-    }
-
-    if (completed === stat.length && typeof callback === 'function') {
-      callback(null, sentBytes);
-    }
+    };
   }
 
-  if (Array.isArray(stat)) {
-    stat.forEach(item => {
-      self.sendStat(item, value, type, sampleRate, tags, onSend);
-    });
-  } else {
-    this.sendStat(stat, value, type, sampleRate, tags, callback);
+  /**
+   * Decorates an async function with timing recording behaviour.
+   *
+   * This version of `timer` will record the time take for the asyncronus action returned by `func`
+   * not just the execution time of `func` itself.
+   *
+   * @param func {<T,A>(...A):Promise<T>} The function to run
+   * @param stat {String|Array} The stat(s) to send
+   * @param sampleRate {Number=} The Number of times to sample (0 to 1). Optional.
+   * @param tags {Array=} The Array of tags to add to metrics. Optional.
+   * @param callback {Function=} Callback when message is done being delivered. Optional.
+   */
+  asyncTimer(func, stat, sampleRate, tags, callback) {
+    return (...args) => {
+      const end = helpers.createHrTimer();
+      const p = func(...args);
+      const recordStat = () => { this.timing(stat, end(), sampleRate, tags, callback); };
+      p.then(recordStat, recordStat);
+      return p;
+    };
   }
-};
 
-/**
- * Sends a stat across the wire
- * @param stat {String|Array} The stat(s) to send
- * @param value The value to send
- * @param type {String} The type of message to send to statsd
- * @param sampleRate {Number} The Number of times to sample (0 to 1)
- * @param tags {Array} The Array of tags to add to metrics
- * @param callback {Function=} Callback when message is done being delivered. Optional.
- */
-Client.prototype.sendStat = function (stat, value, type, sampleRate, tags, callback) {
-  let message = `${this.prefix + stat + this.suffix}:${value}|${type}`;
-  sampleRate = sampleRate || this.sampleRate;
-  if (sampleRate && sampleRate < 1) {
-    if (Math.random() < sampleRate) {
-      message += `|@${sampleRate}`;
-    } else {
-      // don't want to send if we don't meet the sample ratio
-      return callback ? callback() : undefined;
+  /**
+   * Increments a stat by a specified amount
+   * @param stat {String|Array} The stat(s) to send
+   * @param value The value to send
+   * @param sampleRate {Number=} The Number of times to sample (0 to 1). Optional.
+   * @param tags {Array=} The Array of tags to add to metrics. Optional.
+   * @param callback {Function=} Callback when message is done being delivered. Optional.
+   */
+  increment(stat, value, sampleRate, tags, callback) {
+    // allow use of tags without explicit value or sampleRate
+    if (arguments.length < 3) {
+      if (typeof value !== 'number') {
+        tags = value;
+        value = undefined;
+      }
     }
-  }
-  this.send(message, tags, callback);
-};
 
-/**
- * Send a stat or event across the wire
- * @param message {String} The constructed message without tags
- * @param tags {Array} The tags to include (along with global tags). Optional.
- * @param callback {Function=} Callback when message is done being delivered (only if maxBufferSize == 0). Optional.
- */
-Client.prototype.send = function (message, tags, callback) {
-  let mergedTags = this.globalTags;
-  if (tags && typeof tags === 'object') {
-    mergedTags = helpers.overrideTags(mergedTags, tags, this.telegraf);
+    // we explicitly check for undefined and null (and don't do a "! value" check)
+    // so that 0 values are allowed and sent through as-is
+    if (value === undefined || value === null) {
+      value = 1;
+    }
+    this.sendAll(stat, value, 'c', sampleRate, tags, callback);
   }
-  if (mergedTags.length > 0) {
+
+  /**
+   * Decrements a stat by a specified amount
+   * @param stat {String|Array} The stat(s) to send
+   * @param value The value to send
+   * @param sampleRate {Number=} The Number of times to sample (0 to 1). Optional.
+   * @param tags {Array=} The Array of tags to add to metrics. Optional.
+   * @param callback {Function=} Callback when message is done being delivered. Optional.
+   */
+  decrement(stat, value, sampleRate, tags, callback) {
+    this.sendAll(stat, -value || -1, 'c', sampleRate, tags, callback);
+  }
+
+  /**
+   * Represents the histogram stat
+   * @param stat {String|Array} The stat(s) to send
+   * @param value The value to send
+   * @param sampleRate {Number=} The Number of times to sample (0 to 1). Optional.
+   * @param tags {Array=} The Array of tags to add to metrics. Optional.
+   * @param callback {Function=} Callback when message is done being delivered. Optional.
+   */
+  histogram(stat, value, sampleRate, tags, callback) {
+    this.sendAll(stat, value, 'h', sampleRate, tags, callback);
+  }
+
+  /**
+   * Represents the distribution stat
+   * @param stat {String|Array} The stat(s) to send
+   * @param value The value to send
+   * @param sampleRate {Number=} The Number of times to sample (0 to 1). Optional.
+   * @param tags {Array=} The Array of tags to add to metrics. Optional.
+   * @param callback {Function=} Callback when message is done being delivered. Optional.
+   */
+  distribution(stat, value, sampleRate, tags, callback) {
+    this.sendAll(stat, value, 'd', sampleRate, tags, callback);
+  }
+
+  /**
+   * Gauges a stat by a specified amount
+   * @param stat {String|Array} The stat(s) to send
+   * @param value The value to send
+   * @param sampleRate {Number=} The Number of times to sample (0 to 1). Optional.
+   * @param tags {Array=} The Array of tags to add to metrics. Optional.
+   * @param callback {Function=} Callback when message is done being delivered. Optional.
+   */
+  gauge(stat, value, sampleRate, tags, callback) {
+    this.sendAll(stat, value, 'g', sampleRate, tags, callback);
+  }
+
+  /**
+   * Counts unique values by a specified amount
+   * @param stat {String|Array} The stat(s) to send
+   * @param value The value to send
+   * @param sampleRate {Number=} The Number of times to sample (0 to 1). Optional.
+   * @param tags {Array=} The Array of tags to add to metrics. Optional.
+   * @param callback {Function=} Callback when message is done being delivered. Optional.
+   */
+  unique(stat, value, sampleRate, tags, callback) {
+    this.sendAll(stat, value, 's', sampleRate, tags, callback);
+  }
+
+  /**
+   * Counts unique values by a specified amount
+   * @param stat {String|Array} The stat(s) to send
+   * @param value The value to send
+   * @param sampleRate {Number=} The Number of times to sample (0 to 1). Optional.
+   * @param tags {Array=} The Array of tags to add to metrics. Optional.
+   * @param callback {Function=} Callback when message is done being delivered. Optional.
+   */
+  set(stat, value, sampleRate, tags, callback) {
+    this.sendAll(stat, value, 's', sampleRate, tags, callback);
+  }
+
+  /**
+   * Send a service check
+   * @param name {String} The name of the service check
+   * @param status {Number=} The status of the service check (0 to 3).
+   * @param options
+   *   @option date_happened {Date} Assign a timestamp to the event. Default is now.
+   *   @option hostname {String} Assign a hostname to the check.
+   *   @option message {String} Assign a message to the check.
+   * @param tags {Array=} The Array of tags to add to the check. Optional.
+   * @param callback {Function=} Callback when message is done being delivered. Optional.
+   */
+  check(name, status, options, tags, callback) {
     if (this.telegraf) {
-      message = message.split(':');
-      message = `${message[0]},${mergedTags.join(',').replace(/:/g, '=')}:${message.slice(1).join(':')}`;
-    } else {
-      message += `|#${mergedTags.join(',')}`;
+      const err = new Error('Not supported by Telegraf / InfluxDB');
+      if (callback) {
+        return callback(err);
+      } else if (this.errorHandler) {
+        return this.errorHandler(err);
+      }
+
+      throw err;
     }
-  }
 
-  this._send(message, callback);
-};
+    const check = ['_sc', this.prefix + name + this.suffix, status], metadata = options || {};
 
-/**
- * Send a stat or event across the wire
- * @param message {String} The constructed message without tags
- * @param callback {Function=} Callback when message is done being delivered (only if maxBufferSize == 0). Optional.
- */
-Client.prototype._send = function (message, callback) {
-  // we may have a cached error rather than a cached lookup, so
-  // throw it on
-  if (this.dnsError) {
-    if (callback) {
-      return callback(this.dnsError);
-    } else if (this.errorHandler) {
-      return this.errorHandler(this.dnsError);
-    }
-    throw this.dnsError;
-  }
-
-  // Only send this stat if we're not a mock Client.
-  if (!this.mock) {
-    if (this.maxBufferSize === 0) {
-      this.sendMessage(message, callback);
-    } else {
-      this.enqueue(message, callback);
-    }
-  } else {
-    this.mockBuffer.push(message);
-    if (typeof callback === 'function') {
-      callback(null, 0);
-    }
-  }
-};
-
-/**
- * Add the message to the buffer and flush the buffer if needed
- *
- * @param message {String} The constructed message without tags
- */
-Client.prototype.enqueue = function (message, callback) {
-  message += '\n';
-
-  if (this.bufferHolder.buffer.length + message.length > this.maxBufferSize) {
-    this.flushQueue(callback);
-    this.bufferHolder.buffer += message;
-  }
-  else {
-    this.bufferHolder.buffer += message;
-    if (callback) {
-      callback(null);
-    }
-  }
-};
-
-/**
- * Flush the buffer, sending on the messages
- */
-Client.prototype.flushQueue = function (callback) {
-  this.sendMessage(this.bufferHolder.buffer, callback);
-  this.bufferHolder.buffer = '';
-};
-
-/**
- * Send on the message through the socket
- *
- * @param message {String} The constructed message without tags
- * @param callback {Function=} Callback when message is done being delivered. Optional.
- */
-Client.prototype.sendMessage = function (message, callback) {
-  // don't waste the time if we aren't sending anything
-  if (message === '') {
-    if (callback) {
-      callback(null);
-    }
-    return;
-  }
-
-  if (this.protocol === 'tcp' && message.lastIndexOf('\n') !== message.length - 1) {
-    message += '\n';
-  }
-
-  const handleCallback = (err) => {
-    this.messagesInFlight--;
-    const errFormatted = err ? new Error(`Error sending hot-shots message: ${err}`) : null;
-    if (errFormatted) {
-      errFormatted.code = err.code;
-    }
-    if (callback) {
-      callback(errFormatted);
-    } else if (errFormatted) {
-      if (this.errorHandler) {
-        this.errorHandler(errFormatted);
-      } else {
-        console.log(String(errFormatted));
-        // emit error ourselves on the socket for backwards compatibility
-        this.socket.emit('error', errFormatted);
+    if (metadata.date_happened) {
+      const timestamp = helpers.formatDate(metadata.date_happened);
+      if (timestamp) {
+        check.push(`d:${timestamp}`);
       }
     }
-  };
-
-  const buf = Buffer.from(message);
-  try {
-    this.messagesInFlight++;
-    if (this.protocol === 'tcp') {
-      this.socket.write(buf, 'ascii', handleCallback);
-    } else {
-      this.socket.send(buf, 0, buf.length, this.port, this.host, handleCallback);
+    if (metadata.hostname) {
+      check.push(`h:${metadata.hostname}`);
     }
-  } catch (err) {
-    handleCallback(err);
+
+    let mergedTags = this.globalTags;
+    if (tags && typeof(tags) === 'object') {
+      mergedTags = helpers.overrideTags(mergedTags, tags, this.telegraf);
+    }
+    if (mergedTags.length > 0) {
+      check.push(`#${mergedTags.join(',')}`);
+    }
+
+    // message has to be the last part of a service check
+    if (metadata.message) {
+      check.push(`m:${metadata.message}`);
+    }
+
+    // allow for tags to be omitted and callback to be used in its place
+    if (typeof tags === 'function' && callback === undefined) {
+      callback = tags;
+    }
+
+    const message = check.join('|');
+    // Service checks are unique in that message has to be the last element in
+    // the stat if provided, so we can't append tags like other checks. This
+    // directly calls the `_send` method to avoid appending tags, since we've
+    // already added them.
+    this._send(message, callback);
   }
-};
 
-/**
- * Called every bufferFlushInterval to flush any buffer that is around
- */
-Client.prototype.onBufferFlushInterval = function () {
-  this.flushQueue();
-};
-
-/**
- * Close the underlying socket and stop listening for data on it.
- */
-Client.prototype.close = function (callback) {
-  // stop trying to flush the queue on an interval
-  if (this.intervalHandle) {
-    clearInterval(this.intervalHandle);
-  }
-
-  // flush the queue one last time, if needed
-  this.flushQueue((err) => {
-    if (err) {
+  /**
+   * Send on an event
+   * @param title {String} The title of the event
+   * @param text {String} The description of the event.  Optional- title is used if not given.
+   * @param options
+   *   @option date_happened {Date} Assign a timestamp to the event. Default is now.
+   *   @option hostname {String} Assign a hostname to the event.
+   *   @option aggregation_key {String} Assign an aggregation key to the event, to group it with some others.
+   *   @option priority {String} Can be ‘normal’ or ‘low’. Default is 'normal'.
+   *   @option source_type_name {String} Assign a source type to the event.
+   *   @option alert_type {String} Can be ‘error’, ‘warning’, ‘info’ or ‘success’. Default is 'info'.
+   * @param tags {Array=} The Array of tags to add to metrics. Optional.
+   * @param callback {Function=} Callback when message is done being delivered. Optional.
+   */
+  event(title, text, options, tags, callback) {
+    if (this.telegraf) {
+      const err = new Error('Not supported by Telegraf / InfluxDB');
       if (callback) {
         return callback(err);
       }
-      else {
-        return console.log(err);
+      else if (this.errorHandler) {
+        return this.errorHandler(err);
+      }
+
+      throw err;
+    }
+
+    // Convert to strings
+    let message;
+
+    const msgTitle = String(title ? title : '');
+    let msgText = String(text ? text : msgTitle);
+    // Escape new lines (unescaping is supported by DataDog)
+    msgText = msgText.replace(/\n/g, '\\n');
+
+    // start out the message with the event-specific title and text info
+    message = `_e{${msgTitle.length},${msgText.length}}:${msgTitle}|${msgText}`;
+
+    // add in the event-specific options
+    if (options) {
+      if (options.date_happened) {
+        const timestamp = helpers.formatDate(options.date_happened);
+        if (timestamp) {
+          message += `|d:${timestamp}`;
+        }
+      }
+      if (options.hostname) {
+        message += `|h:${options.hostname}`;
+      }
+      if (options.aggregation_key) {
+        message += `|k:${options.aggregation_key}`;
+      }
+      if (options.priority) {
+        message += `|p:${options.priority}`;
+      }
+      if (options.source_type_name) {
+        message += `|s:${options.source_type_name}`;
+      }
+      if (options.alert_type) {
+        message += `|t:${options.alert_type}`;
       }
     }
 
-    // FIXME: we have entered callback hell, and this whole file is in need of an async rework
-
-    // wait until there are no more messages in flight before really closing the socket
-    let intervalAttempts = 0;
-    const waitForMessages = setInterval(() => {
-      intervalAttempts++;
-      if (intervalAttempts > 10) {
-        console.log('hot-shots could not clear out messages in flight but closing anyways');
-        this.messagesInFlight = 0;
-      }
-      if (this.messagesInFlight <= 0) {
-        clearInterval(waitForMessages);
-        this._close(callback);
-      }
-    }, 50);
-  });
-};
-
-/**
- * Really close the socket and handle any errors related to it
- */
-Client.prototype._close = function (callback) {
-  // error function to use in callback and catch below
-  let handledError = false;
-  const handleErr = (err) => {
-    const errMessage = `Error closing hot-shots socket: ${err}`;
-    if (handledError) {
-      console.log(errMessage);
+    // allow for tags to be omitted and callback to be used in its place
+    if (typeof tags === 'function' && callback === undefined) {
+      callback = tags;
     }
-    else {
-      // The combination of catch and error can lead to some errors
-      // showing up twice.  So we just show one of the errors that occur
-      // on close.
-      handledError = true;
 
-      if (callback) {
-        callback(new Error(errMessage));
-      } else if (this.errorHandler) {
-        this.errorHandler(new Error(errMessage));
-      } else {
-        console.log(errMessage);
-      }
-    }
-  };
-
-  if (this.errorHandler) {
-    this.socket.removeListener('error', this.errorHandler);
+    this.send(message, tags, callback);
   }
 
-  // handle error and close events
-  this.socket.on('error', handleErr);
-  if (callback) {
-    this.socket.on('close', err => {
-      if (! handledError && callback) {
-        callback(err);
-      }
+  /**
+   * Creates a child client that adds prefix, suffix and/or tags to this client. Child client can itself have children.
+   * @param options
+   *   @option prefix      {String}  An optional prefix to assign to each stat name sent
+   *   @option suffix      {String}  An optional suffix to assign to each stat name sent
+   *   @option globalTags {Array=} Optional tags that will be added to every metric
+   */
+  childClient(options) {
+    return new Client({
+      isChild     : true,
+      socket      : this.socket, // Child inherits socket from parent. Parent itself can be a child.
+      // All children and parent share the same buffer via sharing an object (cannot mutate strings)
+      bufferHolder: this.bufferHolder,
+      dnsError    : this.dnsError, // Child inherits an error from parent (if it is there)
+      errorHandler: options.errorHandler || this.errorHandler, // Handler for callback errors
+      host        : this.host,
+      port        : this.port,
+      prefix      : (options.prefix || '') + this.prefix, // Child has its prefix prepended to parent's prefix
+      suffix      : this.suffix + (options.suffix || ''), // Child has its suffix appended to parent's suffix
+      globalize   : false, // Only 'root' client can be global
+      mock        : this.mock,
+      // Append child's tags to parent's tags
+      globalTags  : typeof options.globalTags === 'object' ?
+          helpers.overrideTags(this.globalTags, options.globalTags, this.telegraf) : this.globalTags,
+      maxBufferSize : this.maxBufferSize,
+      bufferFlushInterval: this.bufferFlushInterval,
+      telegraf    : this.telegraf,
+      protocol    : this.protocol
     });
   }
-
-  try {
-    if (this.protocol === 'tcp') {
-      this.socket.destroy();
-    } else {
-      this.socket.close();
-    }
-  } catch (err) {
-    handleErr(err);
-  }
-};
-
-const ChildClient = function (parent, options) {
-  options = options || {};
-  Client.call(this, {
-    isChild     : true,
-    socket      : parent.socket, // Child inherits socket from parent. Parent itself can be a child.
-    // All children and parent share the same buffer via sharing an object (cannot mutate strings)
-    bufferHolder: parent.bufferHolder,
-    dnsError    : parent.dnsError, // Child inherits an error from parent (if it is there)
-    errorHandler: options.errorHandler || parent.errorHandler, // Handler for callback errors
-    host        : parent.host,
-    port        : parent.port,
-    prefix      : (options.prefix || '') + parent.prefix, // Child has its prefix prepended to parent's prefix
-    suffix      : parent.suffix + (options.suffix || ''), // Child has its suffix appended to parent's suffix
-    globalize   : false, // Only 'root' client can be global
-    mock        : parent.mock,
-    // Append child's tags to parent's tags
-    globalTags  : typeof options.globalTags === 'object' ?
-        helpers.overrideTags(parent.globalTags, options.globalTags, parent.telegraf) : parent.globalTags,
-    maxBufferSize : parent.maxBufferSize,
-    bufferFlushInterval: parent.bufferFlushInterval,
-    telegraf    : parent.telegraf,
-    protocol    : parent.protocol
-  });
-};
-util.inherits(ChildClient, Client);
-
-/**
- * Creates a child client that adds prefix, suffix and/or tags to this client. Child client can itself have children.
- * @param options
- *   @option prefix      {String}  An optional prefix to assign to each stat name sent
- *   @option suffix      {String}  An optional suffix to assign to each stat name sent
- *   @option globalTags {Array=} Optional tags that will be added to every metric
- */
-Client.prototype.childClient = function (options) {
-  return new ChildClient(this, options);
-};
+}
 
 exports = module.exports = Client;
 exports.StatsD = Client;
