+++ package.json
@@ -1,6 +1,7 @@
 {
   "name": "hot-shots",
   "description": "Node.js client for StatsD, DogStatsD, and Telegraf",
+  "license": "MIT",
   "version": "6.1.1",
   "author": "Steve Ivy",
   "types": "./types.d.ts",
@@ -29,16 +30,26 @@
     "node": ">=6.0.0"
   },
   "scripts": {
-    "coverage": "nyc --reporter=lcov --reporter=text npm test",
-    "test": "mocha --exit -R spec --timeout 5000 test/*.js",
-    "lint": "eslint lib/**.js test/**.js",
-    "pretest": "npm run lint"
+    "test": "npm run test-unit && npm run test-integration",
+    "test-integration": "jest --colors --collectCoverage tests/integration/",
+    "test-unit": "jest --colors --collectCoverage tests/unit/",
+    "lint": "eslint lib/**/*.js tests/**/*.js",
+    "pretest": "npm run lint",
+    "postinstall": "node scripts/postinstall",
+    "benchmark": "node benchmark/statsd.js"
   },
-  "dependencies": {},
   "devDependencies": {
+    "benchmark": "^2.1.4",
     "eslint": "5.9.x",
-    "mocha": "4.x",
-    "nyc": "11.x"
+    "husky": "^1.3.1",
+    "jest": "^24.3.1"
   },
-  "license": "MIT"
+  "husky": {
+    "hooks": {
+      "pre-commit": "npm run test"
+    }
+  },
+  "optionalDependencies": {
+    "unix-dgram": "^2.0.2"
+  }
 }
