+++ tests/integration/close.int.test.js
