+++ tests/integration/globalTags.int.test.js
