+++ tests/integration/statsFunctions.int.test.js
