+++ types.d.ts
@@ -13,8 +13,10 @@ declare module "hot-shots" {
     isChild?: boolean;
     maxBufferSize?: number;
     mock?: boolean;
+    path?: string;
     port?: number;
     prefix?: string;
+    protocol?: 'tcp' | 'udp' | 'unix_dgram';
     sampleRate?: number;
     socket?: dgram.Socket;
     suffix?: string;
